requirements.txt
python-telegram-bot==20.7
google-auth==2.27.0
google-api-python-client==2.116.0
